Update movement logic, add collision detection
by another piece, or if a piece is intending to move to a position
that has a friendly piece (white cannot take white)

Movement logic for Knights needed to be refined, as the number of
possible moves was incomplete.

There was also a bug found where black pieces were overriding white
pieces when taken, this was because as the chessboard array was
refreshed, pieces with duplicate position properties were overriding
each other. This was fixed by updating the board positions and piece
properties concurrently.

Testing needs to be implemented, just the files were created as
placeholders